refactor: module helper
